requirements.txt
Flask
python-dotenv
Flask-Login
Flask-SQLAlchemy
Werkzeug
Authlib
requests